Added fake params for each application

+++ options.h
@@ -162,8 +162,8 @@ struct CodegenOptions {
 };
 
 struct power_analysis_params {
-  int routing_hop_cost;
-  int memtile_op_cost;
+  double routing_hop_cost;
+  double memtile_op_cost;
   map<string, double> alu_op_energy_costs;
 };
